% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augere.solo-package.R
\docType{package}
\name{augere.solo-package}
\alias{augere.solo}
\alias{augere.solo-package}
\title{augere.solo: Automatic Generation of Single-Cell Analyses}
\description{
Implements pipelines for generating single-cell analysis reports in the augere framework. This uses scrapper to execute routine steps such as quality control, normalization, feature selection, clustering and marker detection. Each pipeline function generates a self-contained Rmarkdown report with all of the steps required to reproduce the analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/augere-bioinfo/augere.solo}
  \item Report bugs at \url{https://github.com/augere-bioinfo/augere.solo/issues}
}

}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} (\href{https://orcid.org/0000-0002-3564-4813}{ORCID})

}
