% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnnotate.R
\name{runAnnotate}
\alias{runAnnotate}
\alias{configureReferenceAnnotation}
\title{Cell type annotation from scRNA-seq data}
\usage{
runAnnotate(
  test,
  references,
  test.assay = 1,
  test.id.field = NULL,
  test.block.field = NULL,
  test.is.lognorm = (test.assay == "logcounts"),
  test.is.ensembl = FALSE,
  test.symbol.field = NULL,
  cluster.field = NULL,
  reduced.dimensions = NULL,
  output.dir = "annotate",
  metadata = NULL,
  author = NULL,
  dry.run = FALSE,
  save.results = TRUE,
  num.threads = 1
)

configureReferenceAnnotation(
  ref,
  ref.label.field,
  ref.marker.method = NULL,
  ref.num.markers = NULL,
  ref.assay = "logcounts",
  ref.id.field = NULL,
  ref.block.field = NULL,
  ref.aggregate = NULL,
  ref.is.lognorm = (ref.assay == "logcounts")
)
}
\arguments{
\item{test}{A \link[SummarizedExperiment]{SummarizedExperiment} object containing cells in the test dataset to be assigned labels.}

\item{references}{A list created by \code{configureReferenceAnnotation}, containing a configuration for a reference dataset.

Alternatively, a list of these lists may be supplied to specify multiple references.
This list may be named, in which case the names will be used to identify each reference in both the report and output of this function.}

\item{test.assay}{Integer or string specifying the assay of \code{test} containing the expression values to use for classification.
This is expected to be raw counts or log-transformed normalized expression values.}

\item{test.id.field}{String specifying the name of the \code{\link[SummarizedExperiment]{rowData}(test)} column containing the common gene identifiers.
These identifiers should be consistent with those that are used in each entry of \code{references}.
If \code{NULL}, the row names are assumed to contain the relevant identifiers.}

\item{test.block.field}{String specifying the name of the \code{\link[SummarizedExperiment]{colData}(test)} column containing the block assignment for each cell.
This is used to ensure that some marker-related diagnostics are not driven by uninteresting differences between blocks.
If \code{NULL}, all cells in \code{test} are assumed to originate from the same block.}

\item{test.is.lognorm}{Boolean indicating whether the assay at \code{test.assay} contains log-normalized values.
If \code{FALSE}, it is assumed to contain counts.}

\item{test.is.ensembl}{Boolean indicating whether \code{rownames(test)} contains Ensembl IDs.
Otherwise, it is assumed to contain gene symbols.
Only relevant for entries of \code{references} where \code{ref} is a \pkg{celldex} reference,
where it is used to choose between Ensembl IDs or gene symbols for the row names.}

\item{test.symbol.field}{String specifying the name of the \code{\link[SummarizedExperiment]{rowData}(test)} column that contains gene symbols.
This used to improve the interpretability of some of the gene-based diagnostics in the report.
If \code{NULL}, no symbols are added.}

\item{cluster.field}{String specifying the column of \code{\link[SummarizedExperiment]{colData}(test)} with cluster assignments for each cell.
If provided, this is used to generate some diagnostics to compare the empirical clusters to the predicted labels.}

\item{reduced.dimensions}{Character vector of reduced dimensions in \code{\link[SingleCellExperiment]{reducedDims}(test)} to plot with the predicted labels in the report.}

\item{output.dir}{String containing the path to an output directory in which to write the Rmarkdown report and save results.}

\item{metadata}{Named list of additional fields to add to each result's metadata.}

\item{author}{Character vector of authors.}

\item{dry.run}{Boolean indicating whether to perform a dry run.
This will write the Rmarkdown report without evaluating it.}

\item{save.results}{Boolean indicating whether the results should also be saved to file.}

\item{num.threads}{Integer specifying the number of threads to use in the various computations.}

\item{ref}{A \link[SummarizedExperiment]{SummarizedExperiment} object containing reference samples with known labels in \code{colData(ref)[[label.field]]}.

Alternatively, a string can be provided containing the name of \pkg{celldex} reference dataset (e.g., \code{"HumanPrimaryCellAtlasData"}).
In such cases, \code{label.field} is typically \code{"label.main"} or \code{"label.fine"}.}

\item{ref.label.field}{String specifying the name of the \code{\link[SummarizedExperiment]{colData}(ref)} column containing the label for each reference sample.
For \pkg{celldex} references, this is usually either \code{"label.main"} or \code{"label.fine"}.}

\item{ref.marker.method}{String specifying the method for choosing the top markers from each pairwise comparison between labels,
see the \code{de.method=} argument in \code{\link[SingleR]{trainSingleR}} for more details.
If \code{NULL}, this defaults to \code{"classic"} for the \pkg{celldex} references and \code{"t"} otherwise.}

\item{ref.num.markers}{Integer specifying the number of markers to use from each pairwise comparison between labels.
See the \code{de.n=} argument in \code{\link[SingleR]{trainSingleR}} for more details.}

\item{ref.assay}{Integer or string specifying the assay of \code{ref} containing the expression values to use for creating references.
This should be a matrix-like object that contains counts or their log-normalized values.}

\item{ref.id.field}{String specifying the name of the \code{\link[SummarizedExperiment]{rowData}(ref)} column containing the common gene identifiers.
These identifiers should be consistent with those that are used in \code{test}.
If \code{NULL}, the row names are assumed to contain the relevant identifiers.}

\item{ref.block.field}{String specifying the name of the \code{\link[SummarizedExperiment]{colData}(ref)} column containing the block assignment for each sample.
This is used to ensure that marker calculations are not driven by uninteresting differences between blocks.
If \code{NULL}, all cells in \code{test} are assumed to originate from the same block.}

\item{ref.aggregate}{Boolean indicating that references should be aggregated inside \code{\link[SingleR]{trainSingleR}}.
This can be set to \code{TRUE} for faster classification of large single-cell references.}

\item{ref.is.lognorm}{Boolean indicating whether the assay at \code{assay} contains log-normalized values.
If \code{FALSE}, it is assumed to contain counts.}
}
\value{
For \code{runAnnotate}, a Rmarkdown report named \code{report.Rmd} is written inside \code{output.dir} that contains the analysis commands.

If \code{dry.run=FALSE}, a list is returned containing:
\itemize{
\item \code{predictions}, a list of length equal to \code{references}.
Each entry is a \link[S4Vectors]{DataFrame} containing the classification results of \code{test} against the corresponding reference.
See \code{\link[SingleR]{classifySingleR}} for more details on the expected columns.
\item \code{combined}, a \link[S4Vectors]{DataFrame} containing the combined classification results across multiple references.
See \code{\link[SingleR]{combineRecomputedResults}} for more details on the expected format.
Only present if \code{references} contains more than one reference.
}

If \code{save.results=TRUE}, the results are saved in a \code{results} directory inside \code{output}.

If \code{dry.run=TRUE}, \code{NULL} is returned.
Only the Rmarkdown report is saved to file.

For \code{configureReferenceAnnotation}, a list of class \code{"reference"} is returned containing the configuration details for each reference.
}
\description{
Annotate cells in a scRNA-seq dataset by computing correlations against reference data with known labels.
}
\examples{
library(scRNAseq)
hESCs <- LaMannoBrainData('human-es')
hESCs <- hESCs[,1:100] # speed up for demonstration.

tmp <- tempfile()
results <- runAnnotate(
    hESCs,
    configureReferenceAnnotation(
        "HumanPrimaryCellAtlasData",
        "label.main"
    ),
    output.dir = tmp
)
results$predictions

}
