% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSolo.R
\name{runSolo}
\alias{runSolo}
\title{Simple analysis of single-cell data}
\usage{
runSolo(
  x,
  rna.experiment = TRUE,
  adt.experiment = NULL,
  subset.factor = NULL,
  subset.levels = NULL,
  qc.mito.seqnames = c("MT", "M", "chrM", "chrMT"),
  qc.mito.regex = NULL,
  qc.igg.regex = "IgG|igg|IGG",
  qc.num.mads = 3,
  qc.filter = TRUE,
  num.hvgs = 2000,
  num.pcs = 25,
  cluster.method = "graph",
  cluster.kmeans.k = 10,
  cluster.graph.method = c("multilevel", "leiden", "walktrap"),
  cluster.graph.num.neighbors = 10,
  cluster.graph.resolution = NULL,
  reduced.dimensions = c("tsne", "umap"),
  tsne.perplexity = 30,
  umap.num.neighbors = 15,
  umap.min.dist = 0.1,
  marker.effect.size = c("cohens.d", "auc", "delta.mean", "delta.detected"),
  marker.summary = c("min.rank", "mean", "median", "min"),
  marker.lfc.threshold = 0,
  assay = 1,
  symbol.field = NULL,
  metadata = NULL,
  output.dir = "simple",
  author = NULL,
  dry.run = FALSE,
  save.results = TRUE,
  num.threads = 1
)
}
\arguments{
\item{x}{A \link[SummarizedExperiment]{SummarizedExperiment} object where each column represents a single cell.
Rows are usually expected to contain genes or antibody-derived tags, see \code{rna.experiment=} and \code{adt.experiment=} for details.}

\item{rna.experiment}{Identity of the experiment containing the RNA data, when \code{x} is a \link[SingleCellExperiment]{SingleCellExperiment}.
If \code{TRUE}, the main experiment is assumed to contain the RNA data (see \code{\link[SingleCellExperiment]{mainExpName}}).
If a string is supplied, it is treated as the name of the alteranative experiment (see \code{\link[SingleCellExperiment]{altExps}}) containing the RNA data.
If \code{FALSE} or \code{NULL}, it is assumed that no RNA data is available.}

\item{adt.experiment}{Identity of the experiment containing the ADT data, when \code{x} is a \link[SingleCellExperiment]{SingleCellExperiment}.
If \code{TRUE}, the main experiment is assumed to contain the ADT data (see \code{\link[SingleCellExperiment]{mainExpName}}).
If a string is supplied, it is treated as the name of the alteranative experiment (see \code{\link[SingleCellExperiment]{altExps}}) containing the ADT data.
If \code{FALSE} or \code{NULL}, it is assumed that no ADT data is available.}

\item{subset.factor}{String specifying the name of the \code{\link[SummarizedExperiment]{colData}(se)} column containing a factor with which to subset the cells.}

\item{subset.levels}{Vector containing the subset of levels to retain in the factor specified by \code{subset.factor}.}

\item{qc.mito.seqnames}{Character vector containing the sequence names of the mitochondrial chromosome.
Only used if \code{rna.experiment=} indicates that RNA data is available and \code{qc.mito.regex=} is not specified.}

\item{qc.mito.regex}{String containing a regular expression to identify mitochondrial genes from the row names.
If \code{symbol.field=} is specified, this is applied to the gene symbols instead.
If \code{NULL}, the mitochondrial genes are identified from the sequence names of \code{\link[SummarizedExperiment]{rowRanges}} matching \code{qc.mito.seqnames}. 
Only used if \code{rna.experiment=} indicates that RNA data is available.}

\item{qc.igg.regex}{String containing a regular expression to identify IgG controls from the row names.
If \code{symbol.field=} is specified, this is applied to the gene symbols instead.
Only used if \code{adt.experiment=} indicates that ADT data is available.}

\item{qc.num.mads}{Integer specifying the number of median absolute deviations (MADs) with which to define a quality control (QC) filtering threshold.
Smaller values increase the stringency of the filter.}

\item{qc.filter}{Boolean indicating whether putative low-quality cells should be removed.
If \code{FALSE}, low-quality cells are identified but not removed prior to further cells.}

\item{num.hvgs}{Integer specifying the number of highly variable genes (HVGs) to retain for downstream analyses.
More HVGs capture more biological signal at the cost of capturing more technical noise and increasing computational work.
Only used if \code{rna.experiment=} indicates that RNA data is available.}

\item{num.pcs}{Integer specifying the number of top principal components (PCs) to retain for downstream analyses.
More PCs capture more biological signal at the cost of capturing more technical noise and increasing computational work.}

\item{cluster.method}{Character vector specifying the clustering methods to run on the top PCs,
namely graph-based clustering (\code{"graph"}) or k-means clustering (\code{"kmeans"}).
Multiple methods may be specified here to generate multiple clusterings, but only the clusters from the first method are used for marker detection.}

\item{cluster.kmeans.k}{Integer specifying the number of clusters to generate from k-means clustering.
Only used if \code{"kmeans"} is in \code{cluster.method}.}

\item{cluster.graph.method}{String naming the community detection method to use in graph-based clustering.
These are roughly equivalent to the functions of the same name from the \pkg{igraph} R package.
(Note that \code{"multilevel"} is a synonym for the Louvain method.)
Only used if \code{"graph"} is in \code{cluster.method}.}

\item{cluster.graph.num.neighbors}{Integer specifying the number of nearest neighbors to use during construction of the shared-nearest neighbor graph. 
Larger values increase graph connectivity and decrease cluster resolution. 
Only used if \code{"graph"} is in \code{cluster.method}.}

\item{cluster.graph.resolution}{Number specifying the resolution to use in the multi-level or Leiden community detection algorithms.
Larger values usually result in a greater number of smaller clusters.
Only used if \code{"graph"} is in \code{cluster.method}.}

\item{reduced.dimensions}{Character vector specifying the dimensionality reduction algorithms to use for visualization.
This can be zero, one or both of \code{"tsne"} (for t-SNEs) and \code{"umap"} (for UMAPs).}

\item{tsne.perplexity}{Number specifying the perplexity to use in t-SNE.
Larger values increase the size of each cell's neighborhood and focus more on global structure. 
Only used if \code{"tsne"} is in \code{reduced.dimensions}.}

\item{umap.num.neighbors}{Integer specifying the number of neighbors to use in the UMAP.
Larger values increase connectivity and focus more on global structure. 
Only used if \code{"umap"} is in \code{reduced.dimensions}.}

\item{umap.min.dist}{Number specifying the minimum distance between points in the UMAP.
Larger values favor a more even distribution of cells throughout the low-dimensional space.
Only used if \code{"umap"} is in \code{reduced.dimensions}.}

\item{marker.effect.size}{String naming the effect size to use when ranking marker genes.
This should be one of:
\itemize{
\item \code{"cohens.d"}: Cohen's d, i.e., the standardized difference in log-expression.
This is analogous to the t-statistic in a two-group t-test and accounts for the variance within each group.
\item \code{"auc"}: AUC, i.e., the area under the curve.
This is closely related to the U statistic in a Wilcoxon rank sum test, and is a more robust/less sensitive alternative Cohen's d.
\item \code{"delta.mean"}: difference in the mean log-expression.
This is an estimate of the log-fold change.
\item \code{"delta.detected"}: difference in the proportion of cells with detected expression.
Large differences correspond to genes that are silent in one group and activated in another.
}
This is combined with \code{marker.summary} to determine the actual statistic for ranking, see \code{?\link[scrapper]{scoreMarkers}}.}

\item{marker.summary}{String specifying the summary statistic to use when ranking marker genes.
This should be one of:
\itemize{
\item \code{"min.rank"}: the minimum rank of each gene across all pairwise comparisons for a particular cluster.
This creates a ranking where the top genes are guaranteed to separate the cluster of interest from every other cluster.
\item \code{"mean"}: the mean effect size of each gene across all pairwise comparisons for a particular cluster.
This creates a ranking where the top genes are upregulated in the cluster of interest against the average of all other clusters.
\item \code{"median"}: the median effect size of each gene across all pairwise comparisons for a particular cluster.
This creates a ranking where the top genes are upregulated in the cluster of interest against most other clusters.
\item \code{"min"}: the minimum effect size of each gene across all pairwise comparisons for a particular cluster.
This creates a ranking where the top genes must be upregulated in the cluster of interest against all other clusters.
}
This is combined with \code{marker.effect} to determine the actual statistic for ranking, see \code{?\link[scrapper]{scoreMarkers}}.}

\item{marker.lfc.threshold}{Non-negative number specifying the log-fold change threshold to test against.
Larger values focus on marker genes with larger log-fold changes at the expense of those with smaller variances.}

\item{assay}{String or integer specifying the assay in \code{x} containing the count matrix.
For \link[SingleCellExperiment]{SingleCellExperiment} objects, the same assay is used for the main and alternative experiments.}

\item{symbol.field}{String specifying the name of the \code{\link[SummarizedExperiment]{rowData}} column that contains gene symbols.
This is added to the various gene-based results, e.g., marker detection tables.
If \code{NULL}, no symbols are added.}

\item{metadata}{Named list of additional fields to add to each result's metadata.}

\item{output.dir}{String containing the path to an output directory in which to write the Rmarkdown report and save results.}

\item{author}{Character vector of authors.}

\item{dry.run}{Boolean indicating whether to perform a dry run.
This will write the Rmarkdown report without evaluating it.}

\item{save.results}{Boolean indicating whether the results should also be saved to file.}

\item{num.threads}{Integer specifying the number of threads to use in the various computations.}
}
\value{
A Rmarkdown report named \code{report.Rmd} is written inside \code{output.dir} that contains the analysis commands.

If \code{dry.run=FALSE}, a list is returned containing:
\itemize{
\item \code{sce}, a \link[SingleCellExperiment]{SingleCellExperiment} object with dimensionality reduction and clustering results.
This may have fewer cells than the input \code{x} if \code{qc.filter=TRUE}.
\item \code{qc.rna}, a \link[S4Vectors]{DataFrame} object with RNA-based QC metrics for each cell in the input \code{x}.
Only returned if \code{rna.experiment=} indicates that RNA data is available.
\item \code{qc.adt}, a \link[S4Vectors]{DataFrame} object with ADT-based QC metrics for each cell in the input \code{x}.
Only returned if \code{adt.experiment=} indicates that ADT data is available.
\item \code{markers.rna}, a named list of \link[S4Vectors]{DataFrame} results containing marker gene statistics for each cluster.
Only returned if \code{rna.experiment=} indicates that RNA data is available.
\item \code{markers.adt}, a named list of \link[S4Vectors]{DataFrame} results containing marker tag statistics for each cluster.
Only returned if \code{adt.experiment=} indicates that ADT data is available.
}

If \code{save.results=TRUE}, the results are saved in a \code{results} directory inside \code{output}.

If \code{dry.run=TRUE}, \code{NULL} is returned.
Only the Rmarkdown report is saved to file.
}
\description{
Simple analysis of scRNA-seq or CITE-seq data, from quality control to clustering and marker gene detection.
}
\examples{
# Getting an example dataset from the scRNAseq package.
library(scRNAseq)
se <- ZeiselBrainData()

# Running the pipeline; omitting the save for speed here.
tmp <- tempfile()
output <- runSolo(se, output=tmp, save.results=FALSE)

# Checking the outputs.
list.files(tmp, recursive=TRUE)
output$sce
output$markers

}
\author{
Aaron Lun
}
